Refactor reading from and writing to json file
